# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/forum/stats": {
      "get": {
        "operationId": "ForumController_getStats",
        "summary": "今日论坛发帖数与热帖 Top 5 看板",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/category": {
      "post": {
        "operationId": "ForumController_createCategory",
        "summary": "板块分类-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "ForumController_updateCategory",
        "summary": "板块分类-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/category/list": {
      "get": {
        "operationId": "ForumController_findAllCategories",
        "summary": "板块分类-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "分类名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态（0正常 1停用）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/category/{ids}": {
      "delete": {
        "operationId": "ForumController_removeCategory",
        "summary": "板块分类-删除",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post": {
      "post": {
        "operationId": "ForumController_createPost",
        "summary": "帖子管理-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "ForumController_updatePost",
        "summary": "帖子管理-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/list": {
      "get": {
        "operationId": "ForumController_findAllPosts",
        "summary": "帖子管理-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "分类板块ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "帖子标题（模糊查询）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isTop",
            "required": false,
            "in": "query",
            "description": "是否置顶（0否 1是）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isEssence",
            "required": false,
            "in": "query",
            "description": "是否精华（0否 1是）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态（0正常 1停用）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excludeReported",
            "required": false,
            "in": "query",
            "description": "是否排除已举报帖子（1是 0否）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "发帖人用户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "auditStatus",
            "required": false,
            "in": "query",
            "description": "审核与申诉状态 (0正常 1已下架 2申诉中 3申诉驳回)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/{id}": {
      "get": {
        "operationId": "ForumController_findOnePost",
        "summary": "帖子管理-详情 (会自动累加阅读量)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/{ids}": {
      "delete": {
        "operationId": "ForumController_removePosts",
        "summary": "帖子管理-删除",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/appeal": {
      "post": {
        "operationId": "ForumController_appealPost",
        "summary": "发帖人提起申诉",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/appeal/handle": {
      "put": {
        "operationId": "ForumController_handleAppeal",
        "summary": "管理员审核/复核帖子申诉",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleAppealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/comment": {
      "post": {
        "operationId": "ForumController_createComment",
        "summary": "回帖管理-发布评论",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/comment/list/{postId}": {
      "get": {
        "operationId": "ForumController_findCommentsByPostId",
        "summary": "回帖管理-获取某帖子的嵌套评论盖楼树",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/comment/{id}": {
      "delete": {
        "operationId": "ForumController_removeComment",
        "summary": "回帖管理-删除评论",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/like/{id}": {
      "post": {
        "operationId": "ForumController_togglePostLike",
        "summary": "帖子点赞 (Toggle)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/collect/{id}": {
      "post": {
        "operationId": "ForumController_togglePostCollect",
        "summary": "帖子收藏 (Toggle)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/comment/like/{id}": {
      "post": {
        "operationId": "ForumController_toggleCommentLike",
        "summary": "评论点赞 (Toggle)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/report": {
      "post": {
        "operationId": "ForumController_reportPost",
        "summary": "帖子举报",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/report/list": {
      "get": {
        "operationId": "ForumController_findReports",
        "summary": "帖子举报-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "postId",
            "required": false,
            "in": "query",
            "description": "帖子ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "处理状态（0未处理 1已处理）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/report/handle": {
      "put": {
        "operationId": "ForumController_handleReport",
        "summary": "帖子举报-处理",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandlePostReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/forum/post/collect/my": {
      "get": {
        "operationId": "ForumController_findMyCollectedPosts",
        "summary": "获取当前用户收藏帖子列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "论坛管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "number",
            "description": "分类板块ID"
          },
          "title": {
            "type": "string",
            "description": "帖子标题"
          },
          "content": {
            "type": "string",
            "description": "帖子正文内容"
          },
          "isTop": {
            "type": "string",
            "description": "是否置顶（0否 1是）"
          },
          "isEssence": {
            "type": "string",
            "description": "是否精华（0否 1是）"
          },
          "status": {
            "type": "string",
            "description": "状态（0正常 1停用）"
          }
        },
        "required": [
          "categoryId",
          "title",
          "content"
        ]
      },
...
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "number",
            "description": "分类板块ID"
          },
          "title": {
            "type": "string",
            "description": "帖子标题"
          },
          "content": {
            "type": "string",
            "description": "帖子正文内容"
          },
          "isTop": {
            "type": "string",
            "description": "是否置顶（0否 1是）"
          },
          "isEssence": {
            "type": "string",
            "description": "是否精华（0否 1是）"
          },
          "status": {
            "type": "string",
            "description": "状态（0正常 1停用）"
          },
          "postId": {
            "type": "number",
            "description": "帖子ID"
          }
        },
        "required": [
          "categoryId",
          "title",
          "content",
          "postId"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "sort": {
            "type": "number",
            "description": "显示顺序"
          },
          "status": {
            "type": "string",
            "description": "状态（0正常 1停用）"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "sort": {
            "type": "number",
            "description": "显示顺序"
          },
          "status": {
            "type": "string",
            "description": "状态（0正常 1停用）"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID"
          }
        },
        "required": [
          "name",
          "categoryId"
        ]
      },
      "AppealPostDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "number",
            "description": "被申诉帖子ID"
          },
          "appealReason": {
            "type": "string",
            "description": "申诉原因说明"
          }
        },
        "required": [
          "postId",
          "appealReason"
        ]
      },
      "HandleAppealDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "number",
            "description": "被申诉帖子ID"
          },
          "status": {
            "type": "string",
            "description": "审批状态（0申诉通过/恢复上架 3申诉驳回）"
          }
        },
        "required": [
          "postId",
          "status"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "number",
            "description": "关联帖子ID"
          },
          "parentId": {
            "type": "number",
            "description": "父级评论ID（二级评论或回复时传）"
          },
          "content": {
            "type": "string",
            "description": "评论正文内容"
          }
        },
        "required": [
          "postId",
          "content"
        ]
      },
      "CreatePostReportDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "number",
            "description": "帖子ID"
          },
          "reason": {
            "type": "string",
            "description": "举报原因"
          },
          "content": {
            "type": "string",
            "description": "举报详细说明"
          }
        },
        "required": [
          "postId",
          "reason"
        ]
      },
      "HandlePostReportDto": {
        "type": "object",
        "properties": {
          "reportId": {
            "type": "number",
            "description": "举报ID"
          },
          "status": {
            "type": "string",
            "description": "处理状态（0未处理 1已处理）"
          }
        },
        "required": [
          "reportId",
          "status"
        ]
      }
    }
  }
}